base snapshot sorting on max width of any comparison in build. Break snapshotSort into helper function
